phase8-application-v3/common/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
psycopg[binary]==3.2.1
pydantic==2.7.4
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
redis==5.0.7
aio-pika==9.4.1
# Observability
prometheus_client==0.20.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
